+++ src/core/packager.py
@@ -1,0 +1,5 @@
+# Copyright (C) 2025 Cooper Wenhua <imgwho@gmail.com>
+# SPDX-License-Identifier: AGPL-3.0
+"""Compatibility wrapper for `cwprep.packager`."""
+
+from cwprep.packager import *  # noqa: F403