+++ src/cwprep/commands/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (C) 2025 Cooper Wenhua <imgwho@gmail.com>
+# SPDX-License-Identifier: AGPL-3.0
+"""Command implementations for the cwprep CLI."""